shrink size of `Transducer`

+++ autd3-driver/src/firmware/operation/pulse_width_encoder.rs
@@ -64,7 +64,7 @@ mod tests {
     use super::*;
     use crate::geometry::tests::create_device;
 
-    const NUM_TRANS_IN_UNIT: usize = 249;
+    const NUM_TRANS_IN_UNIT: u8= 249;
 
     #[test]
     fn test() {
